LayoutTests/media/click-volume-bar-not-pausing-expected.txt

Click on volume bar on a MediaDocument does not pause.


Video frame loaded.
EVENT(playing)
EXPECTED (mediaElement.volume == '1') OK
EXPECTED (mediaElement.paused == 'false') OK
EVENT(volumechange)
EXPECTED (mediaElement.volume != '1') OK
EXPECTED (mediaElement.paused == 'false') OK
END OF TEST
